Creación de 'costom hook'
