+++ portafolio/src/componentes/proyectos/proyectosBanner.js
@@ -1,37 +1,10 @@
-import React, { useEffect, useState } from "react";
 import Proyectos from "./proyectos";
 import "../../scss/componentes/proyectos/proyectos.scss";
 import Loader from "../../bibliotecaUI/loader";
+import useVelocidadConexion from "../../hooks/useVelocidadConexion";
 
 const ProyectosBanner = () => {
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const verificaVelocidadConexion = () => {
-      const conexión =
-        navigator.connection ||
-        navigator.mozConnection ||
-        navigator.webkitConnection;
-
-      if (conexión) {
-        const velocidadMbps = conexión.downlink;
-
-        if (velocidadMbps < 2) {
-          setLoading(true);
-        } else {
-          setLoading(false);
-        }
-      }
-    };
-
-    verificaVelocidadConexion();
-
-    const timer = setTimeout(() => {
-      setLoading(false);
-    }, 2000);
-
-    return () => clearTimeout(timer);
-  }, []);
+  const loading = useVelocidadConexion();
 
   return (
     <section className="proyectos-banner">
